spindle pwm

+++ grbl/spindle_control.h
@@ -48,11 +48,10 @@ uint8_t spindle_get_state();
   void spindle_set_state(uint8_t state, float rpm); 
   
   // Sets spindle PWM quickly for stepper ISR. Also called by spindle_set_state().
-  // NOTE: 328p PWM register is 8-bit.
-  void spindle_set_speed(uint8_t pwm_value);
+  void spindle_set_speed(uint32_t pwm_value);
   
-  // Computes 328p-specific PWM register value for the given RPM for quick updating.
-  uint8_t spindle_compute_pwm_value(float rpm);
+  // Computes PWM register value for the given RPM for quick updating.
+  uint32_t spindle_compute_pwm_value(float rpm);
   
 #else
